% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hshmid.R
\name{hshmid}
\alias{hshmid}
\title{Identify hot spot/hot moment (HSHM) observations}
\usage{
hshmid(x, criteria, thresh, side = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{criteria}{The criteria for identifying HSHM. Must be one of "reduce.skew", 
"reduce.kurtosis" or "ref.normal".}

\item{thresh}{The threshold value used to delineate HSHM observations; see Details.}

\item{side}{The side of the distribution to identify HSHM on; one of "upper", "lower"
or "both". Ignored if}
}
\value{
A logical vector indicating whether the corresponding element in \code{x}
is a HSHM.
}
\description{
Identify hot spot/hot moment (HSHM) observations
}
\details{
Currently, three criteria for identifying HSHM are implemented.
"reduce.skew" iteratively removes the most extreme observations until skewness
is \code{<= thresh}. The lowest observations are removed if \code{x} is
left-skewed; the largest observations if \code{x} is right-skewed. The removed 
observations constitute HSHM. "reduce.kurtosis" removes the most extreme observations
from both sides of the distribution until excess kurtosis is \code{<= thresh}.
"ref.normal" determines the values quantile(s) of a normal distribution corresponding to 
\code{thresh} and identifies as HSHM observations in \code{x} that are more extreme 
than the critical value(s). If \code{side} is "upper" or "lower" a one-tailed approach is used;
if \code{side} is "both" a two-tailed approach is used.
}
\examples{
set.seed(11)
x1 <- rnorm(50)
x1[rank(x2) >=45] <- x1[rank(x2) >= 45]*1.6 #create HSHM obs
hshmobs <- hshmid(x1, criteria="ref.normal", side="upper", thresh=0.95)
}
