% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_kurtosis.R
\name{sample_kurtosis}
\alias{sample_kurtosis}
\title{Compute sample kurtosis}
\usage{
sample_kurtosis(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The sample kurtosis of \code{x}.
}
\description{
Compute sample kurtosis
}
\details{
\code{sample_skewness} will return an error if \code{x} contains NA values; address
missing data before using \code{sample_skewness}.
}
\examples{
x1 <- rnorm(27)
sample_kurtosis(x1)

}
\author{
Jonathan Walter, \email{jawalter@ucdavis.edu}
}
