% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_skewness.R
\name{sample_skewness}
\alias{sample_skewness}
\title{Compute sample skewness}
\usage{
sample_skewness(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The sample skewness of \code{x}.
}
\description{
Compute sample skewness
}
\details{
This function uses the definition of sample skewness by Joanes & Gill (1998),
which was shown to perform well in small samples from non-normal distributions. 
\code{sample_skewness} will return an error if \code{x} contains NA values; address
missing data before using \code{sample_skewness}.
}
\examples{
x1 <- rnorm(27)
sample_skewness(x1)

}
\references{
Joanes, D. N.; Gill, C. A. (1998). "Comparing measures of sample skewness and kurtosis". 
Journal of the Royal Statistical Society, Series D. 47 (1): 183–189. doi:10.1111/1467-9884.00122.
}
\author{
Jonathan Walter, \email{jawalter@ucdavis.edu}
}
